added query page
